fix(Hydration): HydateSchedule method updated to use mongoController object instead of always creating new

+++ Lyzer-BE/API/Services/Concrete/HydrationService.cs
@@ -36,8 +36,18 @@ namespace Lyzer_BE.API.Services.Concrete
             if (response.ScheduleData.ScheduleTable.RaceWeekends.Count > 0)
             {
                 MongoController<RaceWeekendDTO> mongoController = new("Schedules", nextYear);
-                var filterValues = Builders<RaceWeekendDTO>.Filter.Empty;
-                mongoController.DeleteManyFromCollection(filterValues);
+
+                bool collectionExists = await mongoController.DoesCollectionExist(nextYear);
+
+                if (!collectionExists)
+                {
+                    await mongoController.CreateCollection(nextYear);
+                }
+                else
+                {
+                    var filterValues = Builders<RaceWeekendDTO>.Filter.Empty;
+                    mongoController.DeleteManyFromCollection(filterValues);
+                }
 
                 if (response.ScheduleData.ScheduleTable.RaceWeekends.Count != 0)
                 {
@@ -48,7 +58,7 @@ namespace Lyzer_BE.API.Services.Concrete
             return response;
         }
 
-        public async Task<ScheduleDTO> HydrateSchedule(string year)
+        public async Task<ScheduleDTO> HydrateSchedule(string year, MongoController<RaceWeekendDTO>? mongoController = null)
         {
             var options = new RestClientOptions("http://ergast.com/api/f1/");
             var client = new RestClient(options);
@@ -58,17 +68,35 @@ namespace Lyzer_BE.API.Services.Concrete
             //TODO: Move duplicated code into method that can be used by other hydration methods.
             if (response.ScheduleData.ScheduleTable.RaceWeekends.Count > 0)
             {
-                MongoController<RaceWeekendDTO> mongoController = new("Schedules", year);
-                var filterValues = Builders<RaceWeekendDTO>.Filter.Empty;
-                mongoController.DeleteManyFromCollection(filterValues);
+                if (mongoController == null)
+                    mongoController = new("Schedules", year);
+
+                bool collectionExists = await mongoController.DoesCollectionExist(year);
+
+                if (!collectionExists)
+                {
+                    await mongoController.CreateCollection(year);
+                }
+                else
+                {
+                    var filterValues = Builders<RaceWeekendDTO>.Filter.Empty;
+                    mongoController.DeleteManyFromCollection(filterValues);
+                }
 
                 if (response.ScheduleData.ScheduleTable.RaceWeekends.Count != 0)
                 {
                     mongoController.InsertManyIntoCollection(response.ScheduleData.ScheduleTable.RaceWeekends);
                 }
+
+                await mongoController.SetCollection(DateTime.Now.Year.ToString());
             }
 
             return response;
+        }
+
+        public async Task<ScheduleDTO> HydrateSchedule(string year)
+        {
+            return await HydrateSchedule(year, null);
         }
     }
 }
